backend/requirements.txt
coverage==3.7.1
django-coverage==1.2.4
djangorestframework
markdown
django-filter
